+++ CHANGELOG.md
@@ -5,9 +5,13 @@ Version changes are pinned to SDK releases.
 
 ## Unreleased
 
+## [1.16.2]
+
+- Prune orders v2. ([#323](https://github.com/zetamarkets/sdk/pull/323))
+
 ## [1.16.1]
 
-- Remove leverage clamp in getMaxTradeSize([#322](https://github.com/zetamarkets/sdk/pull/322))
+- Remove leverage clamp in getMaxTradeSize. ([#322](https://github.com/zetamarkets/sdk/pull/322))
 
 ## [1.16.0]